# Add new index to optimize filters in docs search.